finished admin section in back-end

+++ src/main/java/com/egov/mvc/data/dao/BlogDao.java
@@ -13,8 +13,8 @@ public interface BlogDao {
 
     void addBlog(Blog blog);
     void editBlog(Blog blog);
-    Blog getBlogById(int id);
-    void deleteBlog(int id);
+    Blog getBlogById(long id);
+    void deleteBlog(long id);
     List getAllBlogs();
     user findBlogUserByName(String name);
 }
